fix: remove microsecond rounding with nanos when inserting timestamp
Currently if an OffsetDateTime or LocalTime have > 500 nanoseconds then
TimestampUtils will round up the time's microseconds before inserting
into the database. This causes some problems (e.g. when rounding something
like 9999-12-31 23:59:59.999999+00 to 10000-01-01 00:00:00+00) and is
probably not the expected behavior, which is probably to just truncate the
time and remove the nanoseconds since PostgreSQL doesn't support them. This
change will truncate instead of rounding, and also removes a test of the
removed rounding functionality.

Closes #1412. Could be a breaking change if someone depended upon this
microsecond rounding.